refactor ZooHeader
