+++ packages/lib-react-components/src/ZooHeader/ZooHeader.js
@@ -1,23 +1,16 @@
 import zooTheme from '@zooniverse/grommet-theme'
-import counterpart from 'counterpart'
 import { Anchor, Box } from 'grommet'
 import PropTypes from 'prop-types'
 import React from 'react'
 import styled from 'styled-components'
+import { getHost } from './helpers'
+import { useTranslation } from 'react-i18next'
+import '../translations/i18n'
 
-import en from './locales/en'
 import MainNavList from './components/MainNavList'
 import SignedInUserNavigation from './components/SignedInUserNavigation'
 import SignedOutUserNavigation from './components/SignedOutUserNavigation'
 import ZooniverseLogo from '../ZooniverseLogo'
-import {
-  adminNavLinkLabel,
-  adminNavLinkURL,
-  mainHeaderNavListLabels,
-  mainHeaderNavListURLs
-} from './helpers'
-
-counterpart.registerTranslations('en', en)
 
 export const StyledHeader = styled(Box)`
   color: #B2B2B2;
@@ -41,12 +34,8 @@ export const StyledLogoAnchor = styled(Anchor)`
 
 export default function ZooHeader (props) {
   const {
-    adminNavLinkLabel,
-    adminNavLinkURL,
     isAdmin,
     isNarrow,
-    mainHeaderNavListLabels,
-    mainHeaderNavListURLs,
     register,
     signIn,
     signOut,
@@ -54,6 +43,28 @@ export default function ZooHeader (props) {
     unreadNotifications,
     user
   } = props
+
+  const { t } = useTranslation()
+
+  const host = getHost()
+  const adminNavLinkLabel = 'Admin'
+  const adminNavLinkURL = `${host}/admin`
+
+  const mainHeaderNavListURLs = [
+    `${host}/projects`,
+    `${host}/about`,
+    `${host}/get-involved`,
+    `${host}/talk`,
+    `${host}/lab`
+  ]
+
+  const mainHeaderNavListLabels = [
+    t('ZooHeader.mainHeaderNavListLabels.projects'),
+    t('ZooHeader.mainHeaderNavListLabels.about'),
+    t('ZooHeader.mainHeaderNavListLabels.getInvolved'),
+    t('ZooHeader.mainHeaderNavListLabels.talk'),
+    t('ZooHeader.mainHeaderNavListLabels.build')
+  ]
 
   return (
     <StyledHeader
@@ -69,7 +80,7 @@ export default function ZooHeader (props) {
     >
       <Box
         align='center'
-        aria-label={counterpart('ZooHeader.ariaLabel')}
+        aria-label={t('ZooHeader.ariaLabel')}
         direction='row'
         pad={{ horizontal: 'medium' }}
         responsive={false}
@@ -115,25 +126,17 @@ export default function ZooHeader (props) {
 
 // TODO: remove default prop for register once we add that functionality to auth client
 ZooHeader.defaultProps = {
-  adminNavLinkLabel,
-  adminNavLinkURL,
   breakpoint: 960,
   isAdmin: false,
   isNarrow: false,
-  mainHeaderNavListLabels,
-  mainHeaderNavListURLs,
   register: () => {},
   unreadMessages: 0,
   unreadNotifications: 0
 }
 
 ZooHeader.propTypes = {
-  adminNavLinkLabel: PropTypes.string,
-  adminNavLinkURL: PropTypes.string,
   isAdmin: PropTypes.bool,
   isNarrow: PropTypes.bool,
-  mainHeaderNavListLabels: PropTypes.arrayOf(PropTypes.string),
-  mainHeaderNavListURLs: PropTypes.arrayOf(PropTypes.string),
   register: PropTypes.func,
   signIn: PropTypes.func.isRequired,
   signOut: PropTypes.func.isRequired,
