--- packages/lib-react-components/src/ZooHeader/components/SignedInUserNavigation/locales/en.json
@@ -1,17 +1,0 @@
-{
-  "SignedInUserNavigation": {
-    "ariaLabel": "User account",
-    "navListLabels": {
-      "messages": {
-        "zero": "Messages",
-        "one": "Messages (%(count)s)",
-        "other": "Messages (%(count)s)"
-      },
-      "notifications": {
-        "zero": "Notifications",
-        "one": "Notifications (%(count)s)",
-        "other": "Notifications (%(count)s)"
-      }
-    }
-  }
-}
