--- packages/lib-react-components/src/ZooHeader/components/SignedOutUserNavigation/locales/en.json
@@ -1,6 +1,0 @@
-{
-  "SignedOutUserNavigation": {
-    "register": "Register",
-    "signIn": "Sign In"
-  }
-}
