+++ packages/lib-react-components/src/ZooHeader/helpers/index.js
@@ -1,2 +1,1 @@
 export { default as getHost } from './getHost'
-export { adminNavLinkLabel, adminNavLinkURL, mainHeaderNavListURLs, mainHeaderNavListLabels } from './navListItems'
