--- packages/lib-react-components/src/ZooHeader/locales/en.json
@@ -1,12 +1,0 @@
-{
-  "ZooHeader": {
-    "ariaLabel": "Site",
-    "mainHeaderNavListLabels": {
-      "projects": "Projects",
-      "about": "About",
-      "getInvolved": "Get Involved",
-      "talk": "Talk",
-      "build": "Build"
-    }
-  }
-}
