+++ packages/lib-react-components/src/translations/en.json
@@ -49,5 +49,39 @@
       "talk": "Talk"
     },
     "tagLine": "People-Powered Research"
+  },
+  "ZooHeader": {
+    "ariaLabel": "Site",
+    "mainHeaderNavListLabels": {
+      "projects": "Projects",
+      "about": "About",
+      "getInvolved": "Get Involved",
+      "talk": "Talk",
+      "build": "Build"
+    },
+    "SignedInUserNavigation": {
+      "ariaLabel": "User account",
+      "navListLabels": {
+        "messages_zero": "Messages",
+        "messages_one": "Messages ({{count}})",
+        "messages_other": "Messages ({{count}})",
+        "notifications_zero": "Notifications",
+        "notifications_one": "Notifications ({{count}})",
+        "notifications_other": "Notifications ({{count}})"
+      }
+    },
+    "SignedOutUserNavigation": {
+      "register": "Register",
+      "signIn": "Sign In"
+    },
+    "UserMenu": {
+      "userNavListLabels": {
+        "profile": "Profile",
+        "settings": "Settings",
+        "collections": "Collections",
+        "favorites": "Favorites",
+        "signOut": "Sign out"
+      }
+    }
   }
 }
